+++ docs/CHANGELOG.md
@@ -43,6 +43,7 @@
 - changed the reset keybindings bars appearance to be more visible
 - changed the default exposure bar PC color to blue 2
 - changed lua music `PlayMode` constant names
+- removed the `scripting/trx` directory – internal TRX LUA scripts now get embedded in the exe
 - fixed broken final statistic counters (#4432, regression from 1.0)
 - fixed Bacon Lara not always being drawn perfectly in sync with Lara's animation (#4210)
 - fixed gondolas not being drawn with an underwater tint when they have sunk (#4428)